+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog], and this project adheres to
 [Semantic Versioning].
 
 ## [Unreleased]
+### Fixed
+- order of data types in `DATA VIEWER`.
 
 ------------------------